♻️  streamline tsx => index/...catchAll, part un
