+++ sites/jeromefitzgerald.com/src/components/Layout/Page.tsx
@@ -1,8 +1,39 @@
+import { PageHeading, SkeletonHeading } from '@jeromefitz/design-system/components'
+import { ERROR__FALLBACK } from '@jeromefitz/shared/src/lib/constants'
+import { getNextPageStatus } from 'next-notion/src/utils'
+import dynamic from 'next/dynamic'
+
 import Layout from '~components/Layout'
 import { getRouterNode } from '~routes/index'
 
+const IndexShowListing = dynamic(() => import('~custom/IndexShowListing'), {
+  ssr: false,
+})
+
 const Page = ({ ...props }) => {
-  const { data, dataType, routeType, url } = props
+  const { data, dataType, error, isHomepage, routeType, url } = props
+
+  const { is404, isDataUndefined, isError, isLoading } = getNextPageStatus(
+    data,
+    error,
+    url
+  )
+
+  if (isLoading) return <SkeletonHeading />
+  if (is404)
+    return (
+      <PageHeading
+        description={`Hrm, this page does not seem to want to be found.`}
+        title={'404'}
+      />
+    )
+  if (isError && isDataUndefined)
+    return (
+      <PageHeading
+        description={ERROR__FALLBACK.description}
+        title={ERROR__FALLBACK.title}
+      />
+    )
 
   /**
    * @data
@@ -34,6 +65,7 @@ const Page = ({ ...props }) => {
   return (
     <Layout id={info.id} info={info} routeType={routeType} url={url}>
       <RouterComponent routerNode={routerNode} {...props} />
+      {isHomepage && <IndexShowListing />}
     </Layout>
   )
 }
