+++ sites/jeromefitzgerald.com/src/pages/index.tsx
@@ -1,46 +1,27 @@
-import { PageHeading, SkeletonHeading } from '@jeromefitz/design-system/components'
-import { ERROR__FALLBACK } from '@jeromefitz/shared/src/lib/constants'
 import { nextWeirdRoutingSkipData } from 'next-notion/src/constants'
 import { getCatchAll } from 'next-notion/src/getCatchAll'
 import { getDataReturn } from 'next-notion/src/getDataReturn'
+// import { getStaticPathsCatchAll } from 'next-notion/src/getStaticPathsCatchAll'
 import { getNotion } from 'next-notion/src/helper'
 import { fetcher } from 'next-notion/src/lib/fetcher'
-import { getNextPageStatus } from 'next-notion/src/utils'
-import dynamic from 'next/dynamic'
 import useSWR from 'swr'
 
 import { Page } from '~components/Layout'
 import { notionConfig } from '~config/index'
-// import ShowsListing from '~routes/Shows/Listing'
-
-const IndexShowListing = dynamic(() => import('~custom/IndexShowListing'), {
-  ssr: false,
-})
 
 const notion = getNotion(notionConfig)
 
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
 const { PAGES__HOMEPAGE } = notionConfig
 
-const Index = (props) => {
+const PagesIndex = (props) => {
   const {
     content: contentFallback,
     info: infoFallback,
     images: imagesFallback,
     items: itemsFallback,
-    // hasMeta,
-    // isPage,
-    // isIndex,
-    // meta,
-    // routeType,
     slug,
-    url,
+    // url,
   } = props
-
-  // console.dir(`props`)
-  // console.dir(props)
 
   const { data, error } = useSWR(
     () => (!!slug ? `/api/v1/cms/${slug}` : null),
@@ -61,31 +42,9 @@ const Index = (props) => {
     }
   )
 
-  const { is404, isDataUndefined, isError, isLoading } = getNextPageStatus(
-    data,
-    error,
-    url
-  )
-  if (isLoading) return <SkeletonHeading />
-  if (is404)
-    return (
-      <PageHeading
-        description={`Hrm, this page does not seem to want to be found`}
-        title={'404'}
-      />
-    )
-  if (isError && isDataUndefined)
-    return (
-      <PageHeading
-        description={ERROR__FALLBACK.description}
-        title={ERROR__FALLBACK.title}
-      />
-    )
-
   return (
     <>
-      <Page data={data} {...props} />
-      <IndexShowListing />
+      <Page data={data} error={error} {...props} />
     </>
   )
 }
@@ -108,8 +67,11 @@ export const getStaticProps = async ({ preview = false, ...props }) => {
     }),
     pathVariables,
   })
+
+  const isHomepage = pathVariables?.slug === PAGES__HOMEPAGE
+
   return {
-    props: { preview, ...data, ...pathVariables, ...props },
+    props: { preview, isHomepage, ...data, ...pathVariables, ...props },
     /**
      * @note(next)
      * On-Demand ISR, do not pass revalidate
@@ -119,4 +81,8 @@ export const getStaticProps = async ({ preview = false, ...props }) => {
   }
 }
 
-export default Index
+// export const getStaticPaths = () => {
+//   return getStaticPathsCatchAll(notionConfig)
+// }
+
+export default PagesIndex
